Making sure Popover has correct position
When Popover gets initially rendered with this.props.open = true, the positioning is incorrect (as the logic never gets run). This is because this.componentDidUpdate() is handling the placement when it renders, but that method doesn't get called on the initial render. (https://facebook.github.io/react/docs/component-specs.html#updating-componentdidupdate) To fix this I added this.componentDidMount().